Adds tools for generating web search provider metadata.
The tools consist of a python script and a resource file template. The script downloads a specific source file from the public Chrome source code repository and parses it to retrieve the search engine provider data for the specified locale. It then writes out any new search engine metadata to an existing file under values/all_search_engines.xml and creates locale specific list of search engines in the file res/<values-lang-rRegion>/search_engines.xml (based on the template in this patch).